fix(scanner): accept fractional market_cap_basic from TradingView
TradingView returns market_cap_basic as a fractional float (e.g.
5445242088563.96). The field was annotated `int | None`, so every such row
failed validation and was then silently discarded by
ScannerResponse.from_api_response's blanket `except Exception: continue`.

Effect: any scan selecting market_cap_basic returned zero rows while
reporting a non-zero totalCount. Measured against the live API with
ColumnSets.DETAILED, 25 rows requested:

  america    0/25 parsed    uk       0/25 parsed
  thailand  25/25 parsed    japan   25/25 parsed

US and UK quote fractional caps; TH and JP happened to return integral
values, which is why the breakage went unnoticed. market_cap_basic is the
only affected field.

Widening the annotation to `float | None` restores row parsing
(america: 0 -> 25). Note this changes the exported polars/CSV dtype for
market_cap_basic from Int64 to Float64.

The blanket except that converted a validation error into silent data loss
is left as-is — narrowing it is a behaviour change deserving its own review.

+++ tvkit/api/scanner/models/scanner.py
@@ -234,7 +234,10 @@ class StockData(BaseModel):
     relative_volume_10d_calc: float | None = Field(
         default=None, description="Relative volume compared to 10-day average"
     )
-    market_cap_basic: int | None = Field(default=None, description="Basic market capitalization")
+    # float, not int: TradingView returns fractional market caps (e.g. 5445242088563.96).
+    # A narrower int annotation makes every such row fail validation and be silently
+    # dropped by ScannerResponse.from_api_response.
+    market_cap_basic: float | None = Field(default=None, description="Basic market capitalization")
     fundamental_currency_code: str | None = Field(
         default=None, description="Fundamental currency code"
     )
